Moved authCheck function to util library
