+++ lib/dashboard/index.js
@@ -8,7 +8,8 @@ var express = require('express'),
     expressLess = require('express-less'),
     log = require('../logger'),
     Bundles = require('../bundles'),
-    clientIncludes = require('../client_includes');
+    clientIncludes = require('../client_includes'),
+    utils = require('../util');
 
 log.trace('[lib/dashboard/index.js] Adding Express routes');
 app.use('/dashboard', express.static(__dirname + '/public'));
@@ -19,7 +20,7 @@ app.get('/', function(req, res) {
     res.redirect('/dashboard');
 });
 
-app.get('/dashboard', ensureAuthenticated, function(req, res) {
+app.get('/dashboard', utils.authCheck, function(req, res) {
     Bundles.all()
         .then(function (bundles) {
             res.render('public/dashboard.jade', {bundles: bundles, config: filteredConfig});
@@ -29,20 +30,5 @@ app.get('/dashboard', ensureAuthenticated, function(req, res) {
 });
 
 app.use(clientIncludes);
-
-// TODO: Move this to the future extension API
-function ensureAuthenticated(req, res, next) {
-    if (!config.login.enabled) {
-        return next();
-    }
-
-    var allowed = (req.user !== undefined) ? req.user.allowed : false;
-
-    if (req.isAuthenticated() && allowed) {
-        return next();
-    }
-
-    res.redirect('/login');
-}
 
 module.exports = app;
