+++ lib/extension_api/index.js
@@ -6,6 +6,7 @@ var path = require('path');
 var server = require('../../server.js'); // ew gross TODO fix this
 var log = require('../logger');
 var config = require('../config').filteredConfig;
+var utils = require('../util');
 
 function NodeCG(bundleName, ioInstance) {
     this.bundleName = bundleName;
@@ -171,19 +172,7 @@ NodeCG.prototype.destroySyncedVar = function (args) {
 
 NodeCG.prototype.util = {};
 
-NodeCG.prototype.util.authCheck = function (req, res, next) {
-    if (!config.login.enabled) {
-        return next();
-    }
-
-    var allowed = (req.user !== undefined) ? req.user.allowed : false;
-
-    if (req.isAuthenticated() && allowed) {
-        return next();
-    }
-
-    res.redirect('/login');
-};
+NodeCG.prototype.util.authCheck = utils.authCheck;
 
 Object.defineProperty(NodeCG.prototype, 'extensions', {
     get: function() {
