+++ lib/util/index.js
@@ -1,0 +1,17 @@
+'use strict';
+
+var config = require('../config').config;
+
+module.exports.authCheck = function (req, res, next) {
+    if (!config.login.enabled) {
+        return next();
+    }
+
+    var allowed = (req.user !== undefined) ? req.user.allowed : false;
+
+    if (req.isAuthenticated() && allowed) {
+        return next();
+    }
+
+    res.redirect('/login');
+};
